Adds simulated LiDAR scan for SLAM
Implements a simulated LiDAR scan within the occupancy maze simulator.  This scan updates a separate `slam_grid_map` based on the current robot pose and the environment obstacles.  The `slam_grid_map` is published for visualization and use within SLAM algorithms.

The visualization in RViz is also updated to display the new `slam_grid_map` and to hide the original, complete `grid_map`.  Several new display panels are added to visualize additional data relevant to the SLAM implementation.