Adds reset button to RViz panel
This commit introduces a new RViz plugin that provides a button panel.  The panel includes a "Reset" button which publishes an empty message on the "reset" topic, allowing users to trigger a reset action within the occupancy maze simulator from the RViz interface.  The simulator package structure has been modified, and the image path in the README has been corrected accordingly.